refactor: move dashboard summary feature to ambientAi/
